# [SPARK-52333][SS][PYTHON] Squeeze the protocol of retrieving timers for transformWithState in PySpark