… python/pyspark/sql/streaming/proto/StateMessage_pb2.py