@jingz-db 2025-05-28
I might be missing something, but I think your benchmark results listed in the description is to benchmark against timer related operations and also your code change is built around timer:
```
We benchmarked the change with listing 100 timers (PR for benchmarking: https://github.com/apache/spark/pull/50952), and we saw overall performance improvements.
 ==================== SET IMPLICIT KEY latency (micros) ======================
perc:50		perc:95		perc:99		perc:99.9		perc:100
78.250		141.583		184.375		635.792		962743.500

 ==================== SET IMPLICIT KEY latency (micros) ======================
perc:50		perc:95		perc:99		perc:99.9		perc:100
31.250		47.250		75.875		150.000		551557.750
```

Why do we also see this improvements in set implicit key as well?