@HeartSaVioR 2025-05-29
> IIUC, another benefits of this PR is to get rid of the mid layer of PandasDataFrame and directly use Row. 

It is actually not proved - Scala side does not have the way to pass the Row which Python does not need to deserialize, and vice versa. They also need serde. Arrow RecordBatch is no doubt.

So it's really a matter of which approach has more overhead on how many records and how many columns. Unfortunately I don't have a definite answer for this.